requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
opencv-python==4.5.3.56
numpy==1.21.2
dlib==19.22.0
gunicorn==20.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1 